# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: LifeOS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /assignments:
    get:
      operationId: listAssignments
      tags: [assignments]
      summary: List assignments
      parameters:
        - in: query
          name: status
          schema: { type: ["string", "null"] }
        - in: query
          name: semester
          schema: { type: ["integer", "null"] }
      responses:
        "200":
          description: List of assignments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Assignment"
    post:
      operationId: createAssignment
      tags: [assignments]
      summary: Create an assignment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignmentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assignment"
  /assignments/upcoming:
    get:
      operationId: getUpcomingAssignments
      tags: [assignments]
      summary: Get upcoming assignment/exam deadlines for home dashboard
      parameters:
        - in: query
          name: limit
          schema: { type: integer, default: 3 }
      responses:
        "200":
          description: Upcoming deadlines
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Assignment"
  /assignments/{id}:
    patch:
      operationId: updateAssignment
      tags: [assignments]
      summary: Update an assignment
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignmentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assignment"
    delete:
      operationId: deleteAssignment
      tags: [assignments]
      summary: Delete an assignment
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # EXAMS
  /exams:
    get:
      operationId: listExams
      tags: [exams]
      summary: List exams
      parameters:
        - in: query
          name: semester
          schema: { type: ["integer", "null"] }
      responses:
        "200":
          description: List of exams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Exam"
    post:
      operationId: createExam
      tags: [exams]
      summary: Create an exam
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExamInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exam"
  /exams/{id}:
    patch:
      operationId: updateExam
      tags: [exams]
      summary: Update an exam
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExamUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exam"
    delete:
      operationId: deleteExam
      tags: [exams]
      summary: Delete an exam
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ACHIEVEMENTS
  /achievements:
    get:
      operationId: listAchievements
      tags: [achievements]
      summary: List achievements
      responses:
        "200":
          description: List of achievements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Achievement"
    post:
      operationId: createAchievement
      tags: [achievements]
      summary: Create an achievement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AchievementInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Achievement"
  /achievements/{id}:
    patch:
      operationId: updateAchievement
      tags: [achievements]
      summary: Update an achievement
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AchievementUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Achievement"
    delete:
      operationId: deleteAchievement
      tags: [achievements]
      summary: Delete an achievement
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # PROJECTS
...
  /bucketlist:
    get:
      operationId: listBucketListItems
      tags: [bucketlist]
      summary: List bucket list items
      responses:
        "200":
          description: Bucket list items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BucketListItem"
    post:
      operationId: createBucketListItem
      tags: [bucketlist]
      summary: Create a bucket list item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BucketListItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BucketListItem"
  /bucketlist/{id}:
    patch:
      operationId: updateBucketListItem
      tags: [bucketlist]
      summary: Update a bucket list item
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BucketListItemUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BucketListItem"
    delete:
      operationId: deleteBucketListItem
      tags: [bucketlist]
      summary: Delete a bucket list item
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # FRIENDS
  /friends:
    get:
      operationId: listFriends
      tags: [friends]
      summary: List friends
      responses:
        "200":
          description: List of friends
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Friend"
    post:
      operationId: createFriend
      tags: [friends]
      summary: Create a friend record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FriendInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Friend"
  /friends/{id}:
    patch:
      operationId: updateFriend
      tags: [friends]
      summary: Update a friend
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FriendUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Friend"
    delete:
      operationId: deleteFriend
      tags: [friends]
      summary: Delete a friend record
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # JOURNAL
...
  /future-letters:
    get:
      operationId: listFutureLetters
      tags: [futureLetters]
      summary: List future letters
      responses:
        "200":
          description: Future letters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FutureLetter"
    post:
      operationId: createFutureLetter
      tags: [futureLetters]
      summary: Write a future letter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FutureLetterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FutureLetter"
  /future-letters/{id}:
    delete:
      operationId: deleteFutureLetter
      tags: [futureLetters]
      summary: Delete a future letter
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # STATS
...
components:
  schemas:
...
    Assignment:
      type: object
      required: [id, title, dueDate, status, createdAt]
      properties:
        id: { type: integer }
        title: { type: string }
        subjectId: { type: ["integer", "null"] }
        subjectName: { type: ["string", "null"] }
        dueDate: { type: string }
        priority: { type: string }
        status: { type: string }
        semester: { type: ["integer", "null"] }
        createdAt: { type: string }
    AssignmentInput:
      type: object
      required: [title, dueDate]
      properties:
        title: { type: string }
        subjectId: { type: integer }
        dueDate: { type: string }
        priority: { type: string }
        status: { type: string }
        semester: { type: integer }
    AssignmentUpdate:
      type: object
      properties:
        title: { type: string }
        subjectId: { type: integer }
        dueDate: { type: string }
        priority: { type: string }
        status: { type: string }
        semester: { type: integer }
    Exam:
      type: object
      required: [id, title, examDate, category, createdAt]
      properties:
        id: { type: integer }
        title: { type: string }
        subjectId: { type: ["integer", "null"] }
        subjectName: { type: ["string", "null"] }
        examDate: { type: string }
        category: { type: string }
        semester: { type: ["integer", "null"] }
        grade: { type: ["string", "null"] }
        gradePoint: { type: ["number", "null"] }
        createdAt: { type: string }
    ExamInput:
      type: object
      required: [title, examDate, category]
      properties:
        title: { type: string }
        subjectId: { type: integer }
        examDate: { type: string }
        category: { type: string }
        semester: { type: integer }
        grade: { type: string }
        gradePoint: { type: number }
    ExamUpdate:
      type: object
      properties:
        title: { type: string }
        subjectId: { type: integer }
        examDate: { type: string }
        category: { type: string }
        semester: { type: integer }
        grade: { type: string }
        gradePoint: { type: number }
    Achievement:
      type: object
      required: [id, title, date, createdAt]
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: ["string", "null"] }
        date: { type: string }
        icon: { type: ["string", "null"] }
        category: { type: ["string", "null"] }
        createdAt: { type: string }
    AchievementInput:
      type: object
      required: [title, date]
      properties:
        title: { type: string }
        description: { type: string }
        date: { type: string }
        icon: { type: string }
        category: { type: string }
    AchievementUpdate:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        date: { type: string }
        icon: { type: string }
        category: { type: string }
...
    BucketListItem:
      type: object
      required: [id, text, checked, createdAt]
      properties:
        id: { type: integer }
        text: { type: string }
        checked: { type: boolean }
        targetDate: { type: ["string", "null"] }
        createdAt: { type: string }
    BucketListItemInput:
      type: object
      required: [text]
      properties:
        text: { type: string }
        targetDate: { type: string }
    BucketListItemUpdate:
      type: object
      properties:
        text: { type: string }
        checked: { type: boolean }
        targetDate: { type: string }
    Friend:
      type: object
      required: [id, name, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        department: { type: ["string", "null"] }
        birthday: { type: ["string", "null"] }
        firstMet: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        photoUrl: { type: ["string", "null"] }
        createdAt: { type: string }
    FriendInput:
      type: object
      required: [name]
      properties:
        name: { type: string }
        department: { type: string }
        birthday: { type: string }
        firstMet: { type: string }
        notes: { type: string }
        photoUrl: { type: string }
    FriendUpdate:
      type: object
      properties:
        name: { type: string }
        department: { type: string }
        birthday: { type: string }
        firstMet: { type: string }
        notes: { type: string }
        photoUrl: { type: string }
...
    FutureLetter:
      type: object
      required: [id, title, unlockDate, createdAt]
      properties:
        id: { type: integer }
        title: { type: string }
        content: { type: ["string", "null"] }
        unlockDate: { type: string }
        isUnlocked: { type: boolean }
        createdAt: { type: string }
    FutureLetterInput:
      type: object
      required: [title, unlockDate, content]
      properties:
        title: { type: string }
        content: { type: string }
        unlockDate: { type: string }
...