# lib/api-zod/src/generated/types/bucketListItemInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LifeOS API
 * OpenAPI spec version: 0.1.0
 */

export interface BucketListItemInput {
  text: string;
  targetDate?: string;
}