# lib/api-zod/src/generated/types/examUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LifeOS API
 * OpenAPI spec version: 0.1.0
 */

export interface ExamUpdate {
  title?: string;
  subjectId?: number;
  examDate?: string;
  category?: string;
  semester?: number;
  grade?: string;
  gradePoint?: number;
}